+++ CHANGE_LOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - GH-17 Add DSL and task support for building with Docker labels applied
+- GH-12 Add task for cleaning built docker images, and hook build/clean into base assemble/clean tasks
 
 ## [0.2.0]
 ### Added